net: socionext: align NETSEC driver with v4+ submission
The v3/v4 upstream submissions of the NETSEC driver made a backward
incompatible change to the DT binding, where the PHY node has been
moved into an intermediate MDIO subnode rather than appearing directly
under the NETSEC controller node. The new driver can deal with both
arrangements, but the old driver will cease to work once systems are
upgraded to a new firmware version that uses the new (and official)
version of the binding.

Therefore, we need to update the driver in erp-noupstream, so the v4.14
installer may still be used regardless of the firmware version, allowing
users to upgrade their firmware without regard for which kernel version
they are using.

NOTE: There was a bug in v4 in the microcode loading routine. A fix
      has been incorporated into this patch.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Graeme Gregory <graeme.gregory@linaro.org>