net: phy: core: don't disable device interrupts in phy_change
Commit c34bc2b5059c515aebc3dde4c34baba6db3fd43f upstream.

If state is not PHY_HALTED I see no need to temporarily disable
interrupts on the device. As long as the current interrupt isn't acked
on the device no new interrupt can happen anyway.

In addition remove a unneeded enabling of interrupts in the state
machine when handling state PHY_CHANGELINK.

Tested on a Odroid-C2 with RTL8211F phy in interrupt mode.

Signed-off-by: Heiner Kallweit <hkallweit1@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[ardb: rebase onto v4.14.32/a2c054a896b8ac]
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Graeme Gregory <graeme.gregory@linaro.org>